+++ openflowplugin-impl/src/test/java/org/opendaylight/openflowplugin/impl/statistics/services/direct/multilayer/NodeConnectorDirectStatisticsServiceTest.java
@@ -16,7 +16,6 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import java.math.BigInteger;
 import java.util.Collections;
 import java.util.List;
 import org.opendaylight.mdsal.common.api.LogicalDatastoreType;
@@ -33,6 +32,8 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.protocol.rev130731
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.protocol.rev130731.multipart.request.multipart.request.body.MultipartRequestPortStatsCase;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.protocol.rev130731.multipart.request.multipart.request.body.multipart.request.port.stats._case.MultipartRequestPortStats;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.port.statistics.rev131214.node.connector.statistics.and.port.number.map.NodeConnectorStatisticsAndPortNumberMap;
+import org.opendaylight.yangtools.yang.common.Uint32;
+import org.opendaylight.yangtools.yang.common.Uint64;
 
 public class NodeConnectorDirectStatisticsServiceTest extends AbstractDirectStatisticsServiceTest {
     private PortDirectStatisticsService service;
@@ -53,7 +54,7 @@ public class NodeConnectorDirectStatisticsServiceTest extends AbstractDirectStat
         when(input.getNodeConnectorId()).thenReturn(nodeConnectorId);
 
         final MultipartRequestPortStatsCase body = (MultipartRequestPortStatsCase) ((MultipartRequestInput)service
-            .buildRequest(new Xid(42L), input))
+            .buildRequest(new Xid(Uint32.valueOf(42)), input))
             .getMultipartRequestBody();
 
         final MultipartRequestPortStats nodeConnector = body.getMultipartRequestPortStats();
@@ -75,17 +76,17 @@ public class NodeConnectorDirectStatisticsServiceTest extends AbstractDirectStat
         when(reply.getMultipartReplyBody()).thenReturn(nodeConnectorCase);
 
         when(nodeConnectorStat.getPortNo()).thenReturn(PORT_NO);
-        when(nodeConnectorStat.getTxBytes()).thenReturn(BigInteger.ONE);
-        when(nodeConnectorStat.getCollisions()).thenReturn(BigInteger.ONE);
-        when(nodeConnectorStat.getRxBytes()).thenReturn(BigInteger.ONE);
-        when(nodeConnectorStat.getRxCrcErr()).thenReturn(BigInteger.ONE);
-        when(nodeConnectorStat.getRxDropped()).thenReturn(BigInteger.ONE);
-        when(nodeConnectorStat.getRxErrors()).thenReturn(BigInteger.ONE);
-        when(nodeConnectorStat.getRxFrameErr()).thenReturn(BigInteger.ONE);
-        when(nodeConnectorStat.getRxOverErr()).thenReturn(BigInteger.ONE);
-        when(nodeConnectorStat.getRxPackets()).thenReturn(BigInteger.ONE);
-        when(nodeConnectorStat.getTxDropped()).thenReturn(BigInteger.ONE);
-        when(nodeConnectorStat.getTxErrors()).thenReturn(BigInteger.ONE);
+        when(nodeConnectorStat.getTxBytes()).thenReturn(Uint64.ONE);
+        when(nodeConnectorStat.getCollisions()).thenReturn(Uint64.ONE);
+        when(nodeConnectorStat.getRxBytes()).thenReturn(Uint64.ONE);
+        when(nodeConnectorStat.getRxCrcErr()).thenReturn(Uint64.ONE);
+        when(nodeConnectorStat.getRxDropped()).thenReturn(Uint64.ONE);
+        when(nodeConnectorStat.getRxErrors()).thenReturn(Uint64.ONE);
+        when(nodeConnectorStat.getRxFrameErr()).thenReturn(Uint64.ONE);
+        when(nodeConnectorStat.getRxOverErr()).thenReturn(Uint64.ONE);
+        when(nodeConnectorStat.getRxPackets()).thenReturn(Uint64.ONE);
+        when(nodeConnectorStat.getTxDropped()).thenReturn(Uint64.ONE);
+        when(nodeConnectorStat.getTxErrors()).thenReturn(Uint64.ONE);
 
         final GetNodeConnectorStatisticsOutput output = service.buildReply(input, true);
         assertTrue(output.getNodeConnectorStatisticsAndPortNumberMap().size() > 0);
